fix(api): redact binaries in journal commands
